Change the Redis client patch to use prepend
The current Redis patch uses alias_method which can cause conflicts with
other libraries that use prepend to patch the same methods. This change
moves the existing implementation to a separate Implementation module
and uses prepend to patch methods instead of alias_method.

Fixes #1686
